# docs/08-openapi.yaml
...
openapi: 3.1.0
info:
  title: TokenTrail API
  version: 1.0.0
  description: >
    Control-plane REST API for TokenTrail, the open-source AI cost governance and
    usage analytics platform. Gateway (data-plane) routes under /gw are documented
    separately per provider surface.
  license: { name: Apache-2.0, identifier: Apache-2.0 }
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /workspaces:
    post:
      tags: [Workspaces]
      summary: Create a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
                slug: { type: string, pattern: "^[a-z0-9-]{3,50}$" }
      responses:
        "201": { $ref: "#/components/responses/WorkspaceResponse" }
  /workspaces/{ws}:
    parameters: [{ $ref: "#/components/parameters/ws" }]
    get:
      tags: [Workspaces]
      summary: Get workspace
      responses:
        "200": { $ref: "#/components/responses/WorkspaceResponse" }
    patch:
      tags: [Workspaces]
      summary: Update workspace settings (ADMIN)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                settings:
                  type: object
                  properties:
                    failurePolicy: { type: string, enum: [FAIL_OPEN, FAIL_CLOSED] }
                    memberAnalytics: { type: boolean }
      responses:
        "200": { $ref: "#/components/responses/WorkspaceResponse" }
...
  /workspaces/{ws}/projects:
    parameters: [{ $ref: "#/components/parameters/ws" }]
    get:
      tags: [Projects]
      summary: List projects
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [ACTIVE, ARCHIVED] } }
        - { name: teamId, in: query, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/limit" }
        - { $ref: "#/components/parameters/cursor" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - properties:
                      data: { type: array, items: { $ref: "#/components/schemas/Project" } }
    post:
      tags: [Projects]
      summary: Create project (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                slug: { type: string }
                teamId: { type: string, format: uuid }
                description: { type: string }
                tags: { type: array, items: { type: string } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
...
  /workspaces/{ws}/keys:
    parameters: [{ $ref: "#/components/parameters/ws" }]
    post:
      tags: [VirtualKeys]
      summary: Issue a virtual key. Full key is returned exactly once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, name]
              properties:
                projectId: { type: string, format: uuid }
                userId: { type: string, format: uuid, description: "ADMIN may issue for another user; defaults to caller" }
                name: { type: string }
                providerAllowlist: { type: array, items: { $ref: "#/components/schemas/Provider" } }
                modelAllowlist: { type: array, items: { type: string } }
                rpmLimit: { type: integer, minimum: 1 }
                expiresAt: { type: string, format: date-time }
      responses:
        "201":
          description: Created — `key` is shown only in this response.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/VirtualKey"
                  - properties:
                      key: { type: string, example: "tt_live_9xQ2…" }
  /workspaces/{ws}/keys/{keyId}/revoke:
    parameters:
      - { $ref: "#/components/parameters/ws" }
      - { name: keyId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [VirtualKeys]
      summary: Revoke a key (takes effect at gateway within 5 s)
      responses:
        "200": { description: Revoked }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: "JWT or ttp_ personal access token" }
  parameters:
    ws: { name: ws, in: path, required: true, schema: { type: string }, description: Workspace id or slug }
    limit: { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
    cursor: { name: cursor, in: query, schema: { type: string } }
...
  responses:
...
    WorkspaceResponse:
      description: Workspace
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Workspace" }
  schemas:
...
    Page:
      type: object
      properties:
        data: { type: array, items: {} }
        nextCursor: { type: [string, "null"] }
...
    Provider: { type: string, enum: [ANTHROPIC, OPENAI, GEMINI, MINIMAX, OPENROUTER, DEEPSEEK, OLLAMA] }
...
    Workspace:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        settings: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
...
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        teamId: { type: [string, "null"], format: uuid }
        tags: { type: array, items: { type: string } }
        status: { type: string, enum: [ACTIVE, ARCHIVED] }
...
    VirtualKey:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        keyLast4: { type: string }
        projectId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        providerAllowlist: { type: array, items: { $ref: "#/components/schemas/Provider" } }
        status: { type: string, enum: [ACTIVE, REVOKED, EXPIRED] }
        expiresAt: { type: [string, "null"], format: date-time }
        lastUsedAt: { type: [string, "null"], format: date-time }
...